economic_analysis/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
requests>=2.26.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
python-dotenv>=0.19.0
pytest>=6.2.5
black>=21.5b2
pylint>=2.8.2
